refactor(service): implement generic tools for OpenAI and OpenChat
- Extract generic tool definitions into separate functions
- Create OpenTool struct to represent generic tools
- Implement ToOpenAITool and ToOpenChatTool methods for OpenTool
- Update getOpenAIEmbeddingTools and getOpenChatEmbeddingTools to use generic tools
- Add getOpenWebSearchTool for OpenAI
- Refactor tools.go to use new generic tool functions
- Remove redundant tool definitions from openchat.go
